+++ reusables/__init__.py
@@ -5,24 +5,19 @@
 #
 # Copyright (c) 2014-2017 - Chris Griffith - MIT License
 from __future__ import absolute_import
-from reusables.namespace import *
-from reusables.dt import *
-from reusables.log import *
-from reusables.wrappers import *
-from reusables.web import *
-from reusables.shared_variables import *
+
+from reusables.string_manipulation import *
 from reusables.cli import *
-
-from reusables.file_operations.archive import *
-from reusables.file_operations.common_formats import *
-from reusables.file_operations.config import *
-from reusables.file_operations.file_ops import *
-
-from reusables.processes.helpers import *
-from reusables.processes.tasker import *
-
-from reusables.string_manipulation.numbers import *
-from reusables.string_manipulation.strings import *
+from reusables.dt import *
+from reusables.file_operations import *
+from reusables.log import *
+from reusables.namespace import *
+from reusables.tasker import *
+from reusables.process_helpers import *
+from reusables.shared_variables import *
+from reusables.tasker import *
+from reusables.web import *
+from reusables.wrappers import *
 
 __author__ = "Chris Griffith"
 __version__ = "0.9.1"
