--- reusables/string_manipulation/__init__.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: UTF-8 -*-
-#
-# Part of the Reusables package.
-#
-# Copyright (c) 2014-2017 - Chris Griffith - MIT License
