+++ reusables/wrappers.py
@@ -4,6 +4,7 @@
 # Part of the Reusables package.
 #
 # Copyright (c) 2014-2017 - Chris Griffith - MIT License
+from __future__ import absolute_import
 import time
 from threading import Lock
 from functools import wraps
